Moved RichTextArea to wagtailadmin.rich_text.HalloRichTextArea
