+++ wagtail/wagtailadmin/rich_text.py
@@ -1,0 +1,30 @@
+from __future__ import absolute_import, unicode_literals
+
+import json
+
+from django.forms import widgets
+
+from wagtail.utils.widgets import WidgetWithScript
+from wagtail.wagtailadmin.edit_handlers import RichTextFieldPanel
+from wagtail.wagtailcore.rich_text import DbWhitelister, expand_db_html
+
+
+class HalloRichTextArea(WidgetWithScript, widgets.Textarea):
+    def get_panel(self):
+        return RichTextFieldPanel
+
+    def render(self, name, value, attrs=None):
+        if value is None:
+            translated_value = None
+        else:
+            translated_value = expand_db_html(value, for_editor=True)
+        return super(HalloRichTextArea, self).render(name, translated_value, attrs)
+
+    def render_js_init(self, id_, name, value):
+        return "makeRichTextEditable({0});".format(json.dumps(id_))
+
+    def value_from_datadict(self, data, files, name):
+        original_value = super(HalloRichTextArea, self).value_from_datadict(data, files, name)
+        if original_value is None:
+            return None
+        return DbWhitelister.clean(original_value)
